Move stats/edit page out from crosswords directory
